chore: ChartInformation 파일 생성 및 모델 위치 이전 #28
