+++ Huhoe/Huhoe/Presentation/Scene/HuhoeDetail/ViewModel/ChartInformation.swift
@@ -1,0 +1,15 @@
+//
+//  ChartInformation.swift
+//  Huhoe
+//
+//  Created by 임지성 on 2022/05/05.
+//
+
+import Foundation
+
+struct ChartInformation {
+    let price: [Double]
+    let oldestDate: String
+    let latestDate: String
+    let pointX: Double
+}
